+++ draft-packages/zen-navigation-bar/KaizenDraft/ZenNavigationBar/components/Menu.tsx
@@ -7,7 +7,7 @@ import { OffCanvasContext, ZenOffCanvas } from "@kaizen/draft-zen-off-canvas"
 import classNames from "classnames"
 import Media from "react-media"
 import { NavBarContext } from "../context"
-import { MenuProps, NavigationItem } from "../types"
+import { LinkProps, MenuProps, NavigationItem } from "../types"
 import Dropdown from "./Dropdown"
 import Link from "./Link"
 import MenuGroup from "./MenuGroup"
@@ -37,6 +37,13 @@ class Menu extends React.Component<MenuProps, State> {
 
   state = { open: false }
 
+  toolId: string
+
+  constructor(props: MenuProps) {
+    super(props)
+    this.toolId = uuid()
+  }
+
   render() {
     const { hasExtendedNavigation } = this.context
 
@@ -63,7 +70,7 @@ class Menu extends React.Component<MenuProps, State> {
               <OffCanvasContext.Consumer>
                 {({ toggleVisibleMenu }) => (
                   <Link
-                    key={uuid()}
+                    key={this.toolId}
                     text={`${heading}…`}
                     href="#"
                     onClick={() => toggleVisibleMenu(heading)}
@@ -138,7 +145,8 @@ class Menu extends React.Component<MenuProps, State> {
         if ("url" in item) {
           return (
             <Link
-              key={`${item.url}-${uuid()}`}
+              // These keys should not be derived from the array index
+              key={`${item.url}-${index}`}
               text={item.label}
               href={item.url}
               badge={item.badge}
@@ -147,7 +155,8 @@ class Menu extends React.Component<MenuProps, State> {
         } else if ("title" in item) {
           return (
             <MenuGroup
-              key={`${item.title}-${uuid()}`}
+              // These keys should not be derived from the array index
+              key={`${item.title}-${index}`}
               first={index === 0}
               {...item}
               offCanvas