+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.6...@kaizen/draft-guidance-block@2.4.7) (2020-10-29)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.4.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.5...@kaizen/draft-guidance-block@2.4.6) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.4.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.4...@kaizen/draft-guidance-block@2.4.5) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.4.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.3...@kaizen/draft-guidance-block@2.4.4) (2020-10-27)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [2.4.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.2...@kaizen/draft-guidance-block@2.4.3) (2020-10-21)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block