+++ draft-packages/hierarchical-menu/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.22...@kaizen/draft-hierarchical-menu@1.2.23) (2020-10-29)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
+## [1.2.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.21...@kaizen/draft-hierarchical-menu@1.2.22) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
+## [1.2.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.20...@kaizen/draft-hierarchical-menu@1.2.21) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
+## [1.2.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.19...@kaizen/draft-hierarchical-menu@1.2.20) (2020-10-27)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
 ## [1.2.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.18...@kaizen/draft-hierarchical-menu@1.2.19) (2020-10-21)
 
 **Note:** Version bump only for package @kaizen/draft-hierarchical-menu