+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.4...@kaizen/draft-modal@3.2.5) (2020-10-29)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.3...@kaizen/draft-modal@3.2.4) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.2...@kaizen/draft-modal@3.2.3) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.1...@kaizen/draft-modal@3.2.2) (2020-10-27)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [3.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.0...@kaizen/draft-modal@3.2.1) (2020-10-21)
 
 **Note:** Version bump only for package @kaizen/draft-modal